requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
fonttools==4.51.0
html5lib==1.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==10.3.0
pycparser==2.22
pydyf==0.9.0
pyphen==0.14.0
python-docx==1.1.0
python-docx-replace==0.4.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
zopfli==0.2.3
